Module smartnoise_runtime::utilities::noise

Functions

apply_snapping_noise
Apply noise to value according to the Snapping mechanism. Sensitivity is assumed to be 1 in L1 space.

censored_specific_geom
Return sample from a censored Geometric distribution with parameter p=0.5 without calling to sample_bit_prob.

sample_binomial
Sample from the binomial distribution.

sample_bit

sample_bit_prob
Sample a single bit with arbitrary probability of success

sample_gaussian
Generates a draw from a Gaussian(loc, scale) distribution using the MPFR library.

sample_gaussian_truncated
Sample from truncated Gaussian distribution.

sample_geometric_censored
Sample from the censored geometric distribution with parameter "prob" and maximum number of trials "max_trials".

sample_gumbel
Sample noise from the Gumbel Distribution

sample_laplace
Sample from Laplace distribution centered at shift and scaled by scale.

sample_simple_geometric_mechanism
Sample noise according to geometric mechanism

sample_uniform
Returns random sample from Uniform[min,max).

sample_uniform_int
Sample from uniform integers between min and max (inclusive).

sample_uniform_mpfr
Returns random sample from Uniform[min,max) using the MPFR library.

shuffle
Shuffle a vector